Move the adapters' machinery into `lib/preprocessor/`
The first branch of the Less phase: what only a syntax adapter reads moves out of `lib/utils/` into `lib/preprocessor/`, the directory the preprocessor namespaces will share — the `//`-probe (`readsInlineComments`), the reader's and the writer's questions over it (`endsWithInlineComment`, `movesEndIntoInlineComment`, `writesIntoInlineComment`), `searchCopy`, `requiresTrailingSemicolon`, and the selector copy triad. A move and nothing else: every file keeps its text, the core's adapter follows the paths, and the two types a rule still names (`InlineCommentReading`, `InlineComment`) are re-exported from the `Syntax` contract, so no rule and no checker imports out of the new directory.

The plan had the whole machinery move here at once; what stays behind stays for a reason the boundary test states: the span scanners, `blankComments`, `hideFalseInlineComments`, `rewriteInlineComments` and the interpolation family still have consumers in the core — the named carve-outs of the seam, the `get`/`set` pair keeping the copies in step, the guards — and they follow when the SCSS phase resolves those consumers rather than by loosening the boundary now.

Run over the change: `make verify`; `make oracles` against `origin/main`, all six oracles with no row added, removed or changed.
