+++ AGENTS.md
@@ -25,6 +25,7 @@ A Stylelint plugin (`@stylistic/stylelint-plugin`) that restores the stylisti
 - `lib/index.ts` — builds every rule of the registry once per syntax, through `stylelint.createPlugin(addNamespace(name, syntax.namespace), createRule(syntax))`, and exports the array. `addNamespace` prefixes `@stylistic/`, so users write `"@stylistic/color-hex-case"`; a syntax registered beside the core adds its segment, `"@stylistic/scss/color-hex-case"`.
 - `lib/rules/index.ts` — the registry: static imports of every rule's `createRule` plus a default-exported object keyed by kebab-case rule name. **A new rule is not active until it is added here.**
 - `lib/syntaxes/index.ts` — the `Syntax` contract a rule is built over, and the list of the syntaxes registered beside the core, each under a namespace of its own; `lib/syntaxes/css/` is the core's. A syntax's directory holds everything that is its — the adapter, its tests, its README — and nothing under `lib/rules/` or `lib/utils/` imports out of `lib/syntaxes/` beyond the contract and `css/`, which `lib/syntaxes/index.test.ts` enforces.
+- `lib/preprocessor/` — the machinery only a syntax's adapter reads: the `//`-comment probe and the reader's and writer's questions over it, the selector copy triad, `requiresTrailingSemicolon`. Nothing under `lib/rules/` or `lib/utils/` imports out of it; what still has a consumer in the core — the span scanners, the `has*Interpolation` family — stays in `lib/utils/` until the SCSS phase resolves those consumers.
 - `lib/rules/<rule-name>/` — `index.ts` (rule), `index.test.ts`, `README.md` (user docs).
 - `lib/utils/<utilName>/index.ts` — one util per directory. Many rules are thin wrappers over shared checkers: `whitespaceChecker` (the `always`/`never`/`always-single-line`/… engine behind most `*-space-*` and `*-newline-*` rules), plus the `*CommaSpaceChecker`, `*ColonSpaceChecker`, `isStandardSyntax*`, and `has*Interpolation` families. Look for an existing util before writing new traversal logic — and note that a rule reaches the syntax-aware ones (the `isStandardSyntax*` and `has*Interpolation` families, the inline-comment machinery, the `get`/`set` pairs) only through the `Syntax` contract of `lib/syntaxes/index.ts`, never by importing them. A util puts its helpers above what it exports, and a type is exported where it is declared.
 - `lib/regexps.ts` — every regular expression the plugin reads a stylesheet with, each one a named constant under a line saying what it matches. Nothing else in `lib/` spells a pattern inline, so a pattern wanted in a second place is already there under a name, and whatever had to be worked out about it is written once.
