+++ lib/preprocessor/endsWithInlineComment/index.test.ts
