+++ lib/preprocessor/endsWithInlineComment/index.ts
