+++ lib/preprocessor/movesEndIntoInlineComment/index.test.ts
