+++ lib/preprocessor/movesEndIntoInlineComment/index.ts
