+++ lib/preprocessor/searchCopy/index.test.ts
