+++ lib/preprocessor/toSelectorSourceIndex/index.ts
