+++ lib/preprocessor/writesIntoInlineComment/index.test.ts
