+++ lib/preprocessor/writesIntoInlineComment/index.ts
@@ -1,12 +1,12 @@
 import type { Node } from "postcss"
 import type { PostcssResult } from "stylelint"
 
+import { getAtRuleParams } from "../../utils/getAtRuleParams/index.ts"
+import { getDeclarationValue } from "../../utils/getDeclarationValue/index.ts"
+import { hasBlock } from "../../utils/hasBlock/index.ts"
+import { isAtRule, isComment, isDeclaration } from "../../utils/typeGuards/index.ts"
 import { endsWithInlineComment } from "../endsWithInlineComment/index.ts"
-import { getAtRuleParams } from "../getAtRuleParams/index.ts"
-import { getDeclarationValue } from "../getDeclarationValue/index.ts"
-import { hasBlock } from "../hasBlock/index.ts"
 import { inlineCommentReading } from "../readsInlineComments/index.ts"
-import { isAtRule, isComment, isDeclaration } from "../typeGuards/index.ts"
 
 /** A character standing in for the one the fix would write, put on the end of a run a caller has spelled out. `endsWithInlineComment` reads the trailing whitespace of a text as room a write is about to go into, and such a run is the opposite of that, so something has to stand behind it; only this character's not being whitespace matters, since it neither opens anything nor closes anything. */
 const A_WRITTEN_CHARACTER = `;`
