+++ lib/syntaxes/css/index.ts
@@ -1,10 +1,17 @@
 import type { AtRule, Declaration, Root, Rule as PostcssRule } from "postcss"
 
-import { endsWithInlineComment } from "../../utils/endsWithInlineComment/index.ts"
+import { endsWithInlineComment } from "../../preprocessor/endsWithInlineComment/index.ts"
+import { findSelectorInlineComments, type InlineComment } from "../../preprocessor/findSelectorInlineComments/index.ts"
+import { movesEndIntoInlineComment } from "../../preprocessor/movesEndIntoInlineComment/index.ts"
+import { inlineCommentReading, readsInlineComments, syntaxKeepsInlineComments } from "../../preprocessor/readsInlineComments/index.ts"
+import { requiresTrailingSemicolon } from "../../preprocessor/requiresTrailingSemicolon/index.ts"
+import { restoreSelectorInlineComments } from "../../preprocessor/restoreSelectorInlineComments/index.ts"
+import { searchCopy } from "../../preprocessor/searchCopy/index.ts"
+import { toSelectorSourceIndex } from "../../preprocessor/toSelectorSourceIndex/index.ts"
+import { writesIntoInlineComment } from "../../preprocessor/writesIntoInlineComment/index.ts"
 import { findCommentSpans } from "../../utils/findCommentSpans/index.ts"
 import { findInlineCommentSpans } from "../../utils/findInlineCommentSpans/index.ts"
 import { findInterpolationSpans } from "../../utils/findInterpolationSpans/index.ts"
-import { findSelectorInlineComments, type InlineComment } from "../../utils/findSelectorInlineComments/index.ts"
 import { getAtRuleParams } from "../../utils/getAtRuleParams/index.ts"
 import { getDeclarationValue } from "../../utils/getDeclarationValue/index.ts"
 import { getRuleSelector } from "../../utils/getRuleSelector/index.ts"
@@ -17,18 +24,11 @@ import { isStandardSyntaxProperty } from "../../utils/isStandardSyntaxProperty/i
 import { isStandardSyntaxRule } from "../../utils/isStandardSyntaxRule/index.ts"
 import { isStandardSyntaxSelector } from "../../utils/isStandardSyntaxSelector/index.ts"
 import { isStandardSyntaxValue } from "../../utils/isStandardSyntaxValue/index.ts"
-import { movesEndIntoInlineComment } from "../../utils/movesEndIntoInlineComment/index.ts"
 import { nodeSyntax } from "../../utils/nodeSyntax/index.ts"
-import { inlineCommentReading, readsInlineComments, syntaxKeepsInlineComments } from "../../utils/readsInlineComments/index.ts"
-import { requiresTrailingSemicolon } from "../../utils/requiresTrailingSemicolon/index.ts"
-import { restoreSelectorInlineComments } from "../../utils/restoreSelectorInlineComments/index.ts"
-import { searchCopy } from "../../utils/searchCopy/index.ts"
 import { setAtRuleParams } from "../../utils/setAtRuleParams/index.ts"
 import { setDeclarationValue } from "../../utils/setDeclarationValue/index.ts"
 import { setRuleSelector } from "../../utils/setRuleSelector/index.ts"
-import { toSelectorSourceIndex } from "../../utils/toSelectorSourceIndex/index.ts"
 import { isDeclaration, isRule, type SyntaxRaw } from "../../utils/typeGuards/index.ts"
-import { writesIntoInlineComment } from "../../utils/writesIntoInlineComment/index.ts"
 import type { SelectorCopies, Syntax } from "../index.ts"
 
 /** The syntax of the core: plain CSS, which every rule of the plugin is written for. A styled template is the `styled` namespace's to read, so a root carrying that parser's mark is refused; every other root is still accepted, custom syntaxes without a namespace of their own included. */
